remove wheeltools.py, import/require standalone wheeltools
as discussed in https://github.com/pypa/auditwheel/issues/102
