+++ delocate/tests/test_fuse.py
@@ -9,7 +9,7 @@ from ..tools import (cmp_contents, get_archs, zip2dir, dir2zip, back_tick,
                      open_readable)
 from ..fuse import fuse_trees, fuse_wheels
 from ..tmpdirs import InTemporaryDirectory
-from ..wheeltools import rewrite_record
+from wheeltools import rewrite_record
 
 from .pytest_tools import (assert_true, assert_false, assert_raises,
                         assert_equal, assert_not_equal)
@@ -17,7 +17,10 @@ from .pytest_tools import (assert_true, assert_false, assert_raises,
 
 from .test_tools import LIB32, LIB64, LIB64A
 from .test_wheelies import PURE_WHEEL
-from .test_wheeltools import assert_record_equal
+
+
+def assert_record_equal(record_orig, record_new):
+    assert_equal(sorted(record_orig.splitlines()), sorted(record_new.splitlines()))
 
 
 def assert_same_tree(tree1, tree2):
